refactor: split DateTime checker metadata
