+++ src/types/checker/builtin_types/datetime/timezone.rs
@@ -1,0 +1,312 @@
+//! Purpose:
+//! Synthetic `DateTimeZone` constructors, accessors, and introspection methods.
+//!
+//! Called from:
+//! - DateTime declaration injection.
+//!
+//! Key details:
+//! - Introspection methods are injected only when the timezone bridge is available.
+
+use super::*;
+
+/// `DateTimeZone::__construct(string $timezone = "UTC")` — stores the identifier verbatim.
+pub(super) fn datetime_zone_constructor() -> ClassMethod {
+    method(
+        "__construct",
+        vec![(
+            "timezone".to_string(),
+            Some(TypeExpr::Str),
+            Some(Expr::new(ExprKind::StringLiteral("UTC".to_string()), dummy())),
+            false,
+        )],
+        None,
+        vec![assign_this_property(
+            "name",
+            Expr::new(ExprKind::Variable("timezone".to_string()), dummy()),
+        )],
+    )
+}
+
+/// `DateTimeZone::getName(): string` — returns the stored identifier.
+pub(super) fn datetime_zone_get_name() -> ClassMethod {
+    method("getName", Vec::new(), Some(TypeExpr::Str), vec![return_expr(this_property("name"))])
+}
+
+/// `DateTimeZone::getOffset(DateTimeInterface $datetime): int` — UTC offset (seconds) of this zone
+/// at the given instant.
+///
+/// Temporarily applies this zone via `date_default_timezone_set`, reads the offset with the `date()`
+/// `Z` specifier for `$datetime->getTimestamp()` (so it is daylight-saving correct), then restores
+/// the previous default. Returns a positive value east of UTC, negative west.
+pub(super) fn datetime_zone_get_offset() -> ClassMethod {
+    let call = |name: &str, args: Vec<Expr>| {
+        Expr::new(ExprKind::FunctionCall { name: Name::unqualified(name), args }, dummy())
+    };
+    let var = |n: &str| Expr::new(ExprKind::Variable(n.to_string()), dummy());
+    let expr_stmt = |e: Expr| Stmt::new(StmtKind::ExprStmt(e), dummy());
+    // $datetime->getTimestamp()
+    let dt_ts = Expr::new(
+        ExprKind::MethodCall {
+            object: Box::new(var("datetime")),
+            method: "getTimestamp".to_string(),
+            args: Vec::new(),
+        },
+        dummy(),
+    );
+    let z_spec = Expr::new(ExprKind::StringLiteral("Z".to_string()), dummy());
+    method(
+        "getOffset",
+        vec![(
+            "datetime".to_string(),
+            Some(TypeExpr::Named(Name::unqualified("DateTimeInterface"))),
+            None,
+            false,
+        )],
+        Some(TypeExpr::Int),
+        vec![
+            // $__saved = date_default_timezone_get();
+            Stmt::assign("__saved", call("date_default_timezone_get", Vec::new())),
+            // date_default_timezone_set($this->name);
+            expr_stmt(call("date_default_timezone_set", vec![this_property("name")])),
+            // $__off = intval(date("Z", $datetime->getTimestamp()));
+            Stmt::assign("__off", call("intval", vec![call("date", vec![z_spec, dt_ts])])),
+            // date_default_timezone_set($__saved);  (restore the previous default)
+            expr_stmt(call("date_default_timezone_set", vec![var("__saved")])),
+            return_expr(var("__off")),
+        ],
+    )
+}
+
+/// `DateTimeZone::listIdentifiers(int $timezoneGroup = DateTimeZone::ALL, ?string $countryCode = null): array`
+/// — returns the embedded IANA timezone identifier list. The body is a parsed `return [ ... ];`
+/// over the identifiers in `timezone_ids::TIMEZONE_IDENTIFIERS_ARRAY` (captured from PHP).
+///
+/// The `$timezoneGroup`/`$countryCode` filter parameters are declared for signature parity (so
+/// reflection reports PHP's real signature), but the body returns the full unfiltered list: real
+/// calls are desugared by the name resolver to the injected `__elephc_list_identifiers()` free
+/// function (which performs the group/country filter), so this body only runs via reflection
+/// invocation, where filtering is best-effort.
+pub(super) fn datetime_zone_list_identifiers() -> ClassMethod {
+    let src = format!(
+        "<?php\nreturn [{}];\n",
+        super::timezone_ids::TIMEZONE_IDENTIFIERS_ARRAY
+    );
+    let tokens =
+        crate::lexer::tokenize(&src).expect("listIdentifiers body source must tokenize");
+    let body = crate::parser::parse_internal(&tokens).expect("listIdentifiers body source must parse");
+    ClassMethod {
+        name: "listIdentifiers".to_string(),
+        visibility: Visibility::Public,
+        is_static: true,
+        is_abstract: false,
+        is_final: false,
+        has_body: true,
+        params: vec![
+            (
+                "timezoneGroup".to_string(),
+                Some(TypeExpr::Int),
+                // `DateTimeZone::ALL` (2047) as a literal: referencing the class's own constant in a
+                // default triggers a circular-inheritance error, so the literal value is used.
+                Some(Expr::new(ExprKind::IntLiteral(2047), dummy())),
+                false,
+            ),
+            (
+                "countryCode".to_string(),
+                Some(TypeExpr::Nullable(Box::new(TypeExpr::Str))),
+                Some(Expr::new(ExprKind::Null, dummy())),
+                false,
+            ),
+        ],
+        param_attributes: Vec::new(),
+        variadic: None,
+        variadic_by_ref: false,
+        variadic_type: None,
+        return_type: None,
+        by_ref_return: false,
+        body,
+        span: dummy(),
+        attributes: Vec::new(),
+    }
+}
+
+/// Parses a synthetic-method body from elephc-PHP source into statements. Used so
+/// the introspection methods return array literals directly — the only shape
+/// whose element type a synthetic method's inferred (`None`) return resolves to
+/// (a call to a prelude helper would infer as a scalar). Panics on a
+/// tokenize/parse failure, which is a compiler bug in the static source.
+pub(super) fn parse_tz_body(src: &str) -> Vec<Stmt> {
+    let tokens = crate::lexer::tokenize(src).expect("tz method body must tokenize");
+    crate::parser::parse_internal(&tokens).expect("tz method body must parse")
+}
+
+/// `DateTimeZone::getLocation(): array|false` — returns the zone's country code,
+/// latitude, longitude, and comments (or `false` for the few zones without a
+/// location). Calls the `elephc_tz` bridge directly and marshals the tab-joined
+/// result into an array literal so inference resolves the return shape. Only added
+/// to `DateTimeZone` when the introspection prelude is injected.
+pub(super) fn datetime_zone_get_location() -> ClassMethod {
+    method(
+        "getLocation",
+        Vec::new(),
+        Some(TypeExpr::Named(Name::unqualified("mixed"))),
+        parse_tz_body(
+            r#"<?php
+$raw = elephc_tz_location($this->name);
+if ($raw === "") {
+    return false;
+}
+$f = explode("\t", $raw);
+return [
+    "country_code" => $f[0],
+    "latitude" => (float) $f[1],
+    "longitude" => (float) $f[2],
+    "comments" => $f[3],
+];
+"#,
+        ),
+    )
+}
+
+/// `DateTimeZone::getTransitions(int $timestampBegin = PHP_INT_MIN, int $timestampEnd = PHP_INT_MAX): array|false`
+/// — returns the DST transition rows in the window. The defaults reproduce PHP's
+/// full no-arg list: the synthetic first row coincides with the bridge's row 0, so
+/// its precomputed `time` is reused rather than asking `gmdate` to format
+/// `PHP_INT_MIN`.
+pub(super) fn datetime_zone_get_transitions() -> ClassMethod {
+    // PHP's defaults are PHP_INT_MIN/PHP_INT_MAX. They are materialized as integer
+    // literals (a `ConstRef` default is not evaluated when the method is called
+    // with no args), and `i64::MIN` is exactly the bridge's row-0 timestamp, so the
+    // no-arg call reproduces the full transition list.
+    let int_literal = |v: i64| Expr::new(ExprKind::IntLiteral(v), dummy());
+    method(
+        "getTransitions",
+        vec![
+            (
+                "timestampBegin".to_string(),
+                Some(TypeExpr::Int),
+                Some(int_literal(i64::MIN)),
+                false,
+            ),
+            (
+                "timestampEnd".to_string(),
+                Some(TypeExpr::Int),
+                Some(int_literal(i64::MAX)),
+                false,
+            ),
+        ],
+        Some(TypeExpr::Named(Name::unqualified("mixed"))),
+        parse_tz_body(
+            r#"<?php
+$raw = elephc_tz_transitions($this->name);
+if ($raw === "") {
+    return false;
+}
+$lines = explode("\n", $raw);
+$all = [];
+foreach ($lines as $line) {
+    $g = explode("\t", $line);
+    $all[] = [
+        "ts" => (int) $g[0],
+        "offset" => (int) $g[1],
+        "isdst" => $g[2] === "1",
+        "abbr" => $g[3],
+        "time" => $g[4],
+    ];
+}
+$n = count($all);
+$result = [];
+$active = -1;
+for ($i = 0; $i < $n; $i++) {
+    if ($all[$i]["ts"] <= $timestampBegin) {
+        $active = $i;
+    }
+}
+if ($active >= 0) {
+    $a = $all[$active];
+    // (int) unboxes the boxed array element to a plain int so the comparison with
+    // the int param is reliable (a boxed element compared directly mis-evaluates).
+    // $ats <= $timestampBegin by construction; when they are equal (the
+    // PHP_INT_MIN default lands on row 0, or begin hits a transition exactly),
+    // reuse the bridge's ts/time rather than formatting an extreme begin with
+    // gmdate — gmdate(PHP_INT_MIN) exhausts the heap.
+    $ats = (int) $a["ts"];
+    if ($timestampBegin <= $ats) {
+        // begin coincides with this transition (the PHP_INT_MIN default lands on
+        // row 0): the synthetic row IS this row, so reuse it verbatim. This also
+        // avoids rebuilding an array literal carrying a PHP_INT_MIN value, which the
+        // array machinery mishandles.
+        $result[] = $a;
+    } else {
+        $result[] = [
+            "ts" => $timestampBegin,
+            "time" => gmdate("Y-m-d\TH:i:sP", $timestampBegin),
+            "offset" => $a["offset"],
+            "isdst" => $a["isdst"],
+            "abbr" => $a["abbr"],
+        ];
+    }
+}
+for ($i = 0; $i < $n; $i++) {
+    if ($all[$i]["ts"] > $timestampBegin && $all[$i]["ts"] <= $timestampEnd) {
+        $r = $all[$i];
+        $result[] = [
+            "ts" => $r["ts"],
+            "time" => $r["time"],
+            "offset" => $r["offset"],
+            "isdst" => $r["isdst"],
+            "abbr" => $r["abbr"],
+        ];
+    }
+}
+return $result;
+"#,
+        ),
+    )
+}
+
+/// `DateTimeZone::listAbbreviations(): array` — returns PHP's static
+/// abbreviation→offset/DST/zone table. Static method; calls the `elephc_tz` bridge
+/// directly and marshals the result into the nested array literal.
+pub(super) fn datetime_zone_list_abbreviations() -> ClassMethod {
+    ClassMethod {
+        name: "listAbbreviations".to_string(),
+        visibility: Visibility::Public,
+        is_static: true,
+        is_abstract: false,
+        is_final: false,
+        has_body: true,
+        params: Vec::new(),
+        param_attributes: Vec::new(),
+        variadic: None,
+        variadic_by_ref: false,
+        variadic_type: None,
+        return_type: Some(TypeExpr::Named(Name::unqualified("mixed"))),
+        by_ref_return: false,
+        body: parse_tz_body(
+            r#"<?php
+$raw = elephc_tz_abbreviations();
+$lines = explode("\n", $raw);
+$result = [];
+foreach ($lines as $line) {
+    $parts = explode("\t", $line);
+    $abbr = $parts[0];
+    $rows = explode(";", $parts[1]);
+    $arr = [];
+    foreach ($rows as $row) {
+        $c = explode(":", $row);
+        $id = $c[2];
+        $arr[] = [
+            "dst" => $c[0] === "1",
+            "offset" => (int) $c[1],
+            "timezone_id" => ($id === "NULL" ? null : $id),
+        ];
+    }
+    $result[$abbr] = $arr;
+}
+return $result;
+"#,
+        ),
+        span: dummy(),
+        attributes: Vec::new(),
+    }
+}
